+++ src/check_jsonschema/loaders/__init__.py
@@ -1,6 +1,11 @@
 from .errors import BadFileTypeError, SchemaParseError, UnsupportedUrlScheme
 from .instance import InstanceLoader, instance_loader_from_args
-from .schema import BuiltinSchemaLoader, SchemaLoader, schema_loader_from_args
+from .schema import (
+    BuiltinSchemaLoader,
+    SchemaLoader,
+    SchemaLoaderBase,
+    schema_loader_from_args,
+)
 
 __all__ = (
     "BadFileTypeError",
@@ -8,6 +13,7 @@ __all__ = (
     "UnsupportedUrlScheme",
     "BuiltinSchemaLoader",
     "SchemaLoader",
+    "SchemaLoaderBase",
     "InstanceLoader",
     "schema_loader_from_args",
     "instance_loader_from_args",